refactor(attendance/serializers): use a structured folder instead of a chunk of monolithic file
